Split linkerd2-proxy-resolve
`linkerd2-proxy::control` included the logic used to resolve names from
the control plane. Before this logic is altered to support IP-address
names, this change splits this module into a new subcrate with minimal
other changes.

No functional changes.
