+++ lib/linkerd2-proxy-resolve/Cargo.toml
@@ -1,0 +1,19 @@
+[package]
+name = "linkerd2-proxy-resolve"
+version = "0.1.0"
+authors = ["Oliver Gould <ver@buoyant.io>"]
+edition = "2018"
+
+[dependencies]
+futures = "0.1"
+linkerd2-addr = { path = "../linkerd2-addr" }
+linkerd2-dns-name = { path = "../linkerd2-dns-name" }
+linkerd2-identity = { path = "../linkerd2-identity" }
+linkerd2-proxy-api = { git = "https://github.com/linkerd/linkerd2-proxy-api", rev = "ddbc3a4f7f8b0058801f896d27974d19ee98094c" }
+linkerd2-proxy-core = { path = "../linkerd2-proxy-core" }
+linkerd2-never = { path = "../linkerd2-never" }
+linkerd2-task = { path = "../linkerd2-task" }
+prost = "0.5.0"
+tower-grpc = { git = "https://github.com/tower-rs/tower-grpc", default-features = false, features = ["protobuf"] }
+indexmap = "1.0"
+tracing = "0.1"
