+++ lib/linkerd2-proxy-resolve/src/destination/resolution.rs
@@ -1,4 +1,5 @@
 use super::client;
+use crate::addr::NameAddr;
 use crate::api::{
     destination::{
         protocol_hint::Protocol, update::Update as PbUpdate2, TlsIdentity, Update as PbUpdate,
@@ -6,9 +7,9 @@ use crate::api::{
     },
     net::TcpAddress,
 };
-use crate::control::destination::{Metadata, ProtocolHint};
 use crate::core::resolve::{self, Update};
-use crate::{identity, logging, NameAddr, Never};
+use crate::destination::{Metadata, ProtocolHint};
+use crate::{identity, task, Never};
 use futures::{
     future::Future,
     sync::{mpsc, oneshot},
@@ -16,7 +17,6 @@ use futures::{
 };
 use indexmap::{IndexMap, IndexSet};
 use std::{collections::HashMap, error::Error, fmt, net::SocketAddr};
-use tokio;
 use tower_grpc::{self as grpc, generic::client::GrpcService, Body, BoxBody};
 use tracing::{debug, trace, warn};
 
@@ -165,10 +165,9 @@ where
                 .expect("resolution should not have been dropped");
 
             let query = self.query.take().expect("invalid state");
-            let ctx = logging::Section::Proxy.bg(LogCtx(query.authority().clone()));
 
-            let daemon = ctx.future(Daemon { updater, query });
-            tokio::spawn(Box::new(daemon));
+            // todo
+            task::spawn(Daemon { updater, query });
 
             return Ok(Async::Ready(res));
         }
