+++ src/app/proxy.rs
@@ -1,5 +1,5 @@
-use crate::drain;
 use crate::core::listen::{ListenAndSpawn, ServeConnection};
+use crate::drain;
 use futures::{self, Future, Poll};
 use tracing::error;
