--- src/control/mod.rs
@@ -1,6 +1,0 @@
-pub mod destination;
-mod remote_stream;
-mod serve_http;
-
-pub use self::serve_http::serve_http;
-pub use self::serve_http::ClientAddr;
