+++ src/proxy/resolve.rs
@@ -1,5 +1,5 @@
-use crate::{svc, Error};
 use crate::core::resolve::{Resolution, Resolve, Update};
+use crate::{svc, Error};
 use futures::{stream::FuturesUnordered, try_ready, Async, Future, Poll, Stream};
 use indexmap::IndexMap;
 use std::{fmt, net::SocketAddr};
@@ -270,12 +270,14 @@ mod tests {
     use tower_discover::{Change, Discover as _Discover};
     use tower_util::service_fn;
 
-    impl<E> Resolution for mpsc::Receiver<Update<E>> {
+    struct UpdateRx<E>(mpsc::Receiver<Update<E>>);
+
+    impl<E> Resolution for UpdateRx<E> {
         type Endpoint = E;
         type Error = mpsc::error::RecvError;
 
         fn poll(&mut self) -> Poll<Update<Self::Endpoint>, Self::Error> {
-            let ep = try_ready!(Stream::poll(self)).expect("stream must not terminate");
+            let ep = try_ready!(Stream::poll(&mut self.0)).expect("stream must not terminate");
             Ok(Async::Ready(ep))
         }
     }
@@ -304,7 +306,7 @@ mod tests {
             let (make1_tx, make1_rx) = oneshot::channel::<Svc<usize>>();
             let make = Svc(vec![make1_rx, make0_rx]);
 
-            let mut discover = Discover::new(resolution, make);
+            let mut discover = Discover::new(UpdateRx(resolution), make);
             assert!(
                 discover.poll().expect("discover can't fail").is_not_ready(),
                 "ready without updates"
@@ -407,7 +409,7 @@ mod tests {
             let (make1_tx, make1_rx) = oneshot::channel::<Svc<usize>>();
             let make = Svc(vec![make1_rx, make0_rx]);
 
-            let mut discover = Discover::new(resolution, make);
+            let mut discover = Discover::new(UpdateRx(resolution), make);
             assert!(
                 discover.poll().expect("discover can't fail").is_not_ready(),
                 "ready without updates"
@@ -482,7 +484,7 @@ mod tests {
             let (mut tx, resolution) = mpsc::channel(1);
             let make = service_fn(|()| future::empty::<Svc<()>, Error>());
 
-            let mut discover = Discover::new(resolution, make);
+            let mut discover = Discover::new(UpdateRx(resolution), make);
             assert!(
                 discover.poll().expect("discover can't fail").is_not_ready(),
                 "ready without updates"
